pyproject.toml
[project]
name = "trustforge"
version = "0.1.0"
description = "Trustforge: Policy Foundry with tokenized themes (PDF/HTML/CSV)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "CNCISO", email = "info@cloudnativeciso.com" }]
dependencies = [
  "pydantic>=2.8.2",
  "typer>=0.12.5",
  "pyyaml>=6.0.2",
  "markdown-it-py>=3.0.0",
  "jinja2>=3.1.4",
]
[project.scripts]
trustforge = "trustforge.cli:app"

[project.optional-dependencies]
dev = [
  "ruff>=0.6.9",
  "mypy>=1.10.0",
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "types-PyYAML",
  "types-Markdown",
]

[tool.ruff]
line-length = 100
target-version = "py311"
fix = true                   # <— belongs here

[tool.ruff.lint]
extend-select = ["I"]        # import sorting (isort)

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
no_implicit_optional = true

# src/ layout — prevents “Source file found twice …”
explicit_package_bases = true
mypy_path = "src"
files = ["src"]
packages = ["trustforge"]
